+++ src/main/java/com/team9/anicare/animal/repository/BreedRepository.java
@@ -1,7 +1,7 @@
-package com.team9.anicare.species.repository;
+package com.team9.anicare.animal.repository;
 
-import com.team9.anicare.species.model.Breed;
-import com.team9.anicare.species.model.Species;
+import com.team9.anicare.animal.model.Breed;
+import com.team9.anicare.animal.model.Species;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
